+++ Assets/Script/Status.cs
@@ -13,8 +13,8 @@ public class Status : ScriptableObject
     public float HP = 0;
     public float MAXMP = 0;
     public float MP = 0;
-    public float ATK = 0;
-    public float DEF = 0;
+    public int ATK = 0;
+    public int DEF = 0;
 
     public GameObject obj = null;
